test: consolidate and expand filename analysis tests
Merged `tests/test_filename_analysis_extended.py` into `tests/test_filename_analysis.py` to reduce file fragmentation and improve test hygiene.

The consolidated test suite now covers:
- RTLO character detection
- Double extension detection with various separators (spaces, tabs, multiple dots)
- Deceptive prefixes (.svg, .html) and additional executable extensions (.lnk, .inf)
- Whitespace tricks (large gaps, space before single extension)
- Trailing spaces and dots
- Invisible Unicode characters (Zero-width space, Soft hyphen) and control characters
- Integration with the global `Config` class

All 15 test cases in the consolidated file pass, as does the full test suite.
